Revert "Nodejs Runtime as Node server (#443)"
This reverts commit 574591991d2a232623c2015c62f77743d8bc23c1.
