+++ Makefile
@@ -216,17 +216,6 @@ processor-shell: processor
 
 IMAGES_TO_PUSH += $(NUCLIO_DOCKER_PROCESSOR_SHELL_ALPINE_IMAGE_NAME)
 
-# nodejs
-NUCLIO_HANDLER_NODEJS_DOCKERFILE_PATH = pkg/processor/build/runtime/nodejs/docker/Dockerfile.handler-nodejs
-NUCLIO_DOCKER_HANDLER_NODEJS_ALPINE_IMAGE_NAME=nuclio/handler-nodejs:$(NUCLIO_DOCKER_IMAGE_TAG_WITH_ARCH)
-
-handler-nodejs: processor
-	docker build $(NUCLIO_BUILD_ARGS_VERSION_INFO_FILE) \
-	-f $(NUCLIO_HANDLER_NODEJS_DOCKERFILE_PATH) \
-	-t $(NUCLIO_DOCKER_HANDLER_NODEJS_ALPINE_IMAGE_NAME) .
-
-IMAGES_TO_PUSH += $(NUCLIO_DOCKER_HANDLER_NODEJS_ALPINE_IMAGE_NAME)
-
 #
 # Testing
 #
