--- pkg/processor/runtime/nodejs/runtime.go
@@ -1,120 +1,0 @@
-/*
-Copyright 2017 The Nuclio Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package nodejs
-
-import (
-	"fmt"
-	"os"
-	"os/exec"
-	"strings"
-
-	"github.com/nuclio/nuclio/pkg/common"
-	"github.com/nuclio/nuclio/pkg/errors"
-	"github.com/nuclio/nuclio/pkg/processor/runtime"
-	"github.com/nuclio/nuclio/pkg/processor/runtime/rpc"
-
-	"github.com/nuclio/nuclio-sdk"
-)
-
-type nodejs struct {
-	*rpc.Runtime
-	Logger        nuclio.Logger
-	configuration *runtime.Configuration
-}
-
-// NewRuntime returns a new NodeJS runtime
-func NewRuntime(parentLogger nuclio.Logger, configuration *runtime.Configuration) (runtime.Runtime, error) {
-	newNodeJSRuntime := &nodejs{
-		configuration: configuration,
-		Logger:        parentLogger.GetChild("nodejs"),
-	}
-	//newNodeJSRuntime.Logger = logger // We *must* initialize logger here
-
-	var err error
-	newNodeJSRuntime.Runtime, err = rpc.NewRPCRuntime(newNodeJSRuntime.Logger, configuration, newNodeJSRuntime.runWrapper)
-
-	return newNodeJSRuntime, err
-}
-
-// We can't use n.Logger since it's not initialized
-func (n *nodejs) runWrapper(socketPath string) error {
-	wrapperScriptPath := n.getWrapperScriptPath()
-	n.Logger.DebugWith("Using nodejs wrapper script path", "path", wrapperScriptPath)
-	if !common.IsFile(wrapperScriptPath) {
-		return fmt.Errorf("Can't find wrapper at %q", wrapperScriptPath)
-	}
-
-	nodeExePath, err := n.getNodeExePath()
-	if err != nil {
-		n.Logger.ErrorWith("Can't find node exe", "error", err)
-		return errors.Wrap(err, "Can't find node exe")
-	}
-	n.Logger.DebugWith("Using node executable", "path", nodeExePath)
-
-	handlerPath := n.getHandlerPath()
-
-	// pass global environment onto the process, and sprinkle in some added env vars
-	env := os.Environ()
-	env = append(env, n.getEnvFromConfiguration()...)
-	args := []string{nodeExePath, wrapperScriptPath, socketPath, handlerPath}
-
-	n.Logger.DebugWith("Running wrapper", "command", strings.Join(args, " "))
-
-	cmd := exec.Command(args[0], args[1:]...)
-	cmd.Env = env
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stdout
-
-	return cmd.Start()
-}
-
-func (n *nodejs) getEnvFromConfiguration() []string {
-	return []string{
-		fmt.Sprintf("NUCLIO_FUNCTION_NAME=%s", n.configuration.Meta.Name),
-		fmt.Sprintf("NUCLIO_FUNCTION_DESCRIPTION=%s", n.configuration.Spec.Description),
-		fmt.Sprintf("NUCLIO_FUNCTION_VERSION=%d", n.configuration.Spec.Version),
-	}
-}
-
-func (n *nodejs) getHandlerPath() string {
-	handlerPath := os.Getenv("NUCLIO_JS_HANDLER")
-	if handlerPath != "" {
-		return handlerPath
-	}
-
-	return "/opt/nuclio/handler/handler.js"
-}
-
-// TODO: Global processor configuration, where should this go?
-func (n *nodejs) getWrapperScriptPath() string {
-	scriptPath := os.Getenv("NUCLIO_NODEJS_WRAPPER_PATH")
-	if len(scriptPath) == 0 {
-		return "/opt/nuclio/wrapper.js"
-	}
-
-	return scriptPath
-}
-
-func (n *nodejs) getNodeExePath() (string, error) {
-	exePath := os.Getenv("NUCLIO_NODEJS_EXE")
-	if exePath != "" {
-		return exePath, nil
-	}
-	baseName := "node"
-
-	return exec.LookPath(baseName)
-}
