+++ pkg/processor/runtime/python/doc.go
@@ -15,24 +15,19 @@ limitations under the License.
 */
 
 /*
-Package rpc implement RPC runtime
+Package python implement Python runtime
 
-The Go RPC runtime opens a Unix socket and starts a wrapper script with path to
-the socket and the handler to run. The wrapper connects to this socket upon
-startup.
+The Go Python runtime opens a Unix socket and starts the wrapper Python script
+(`wrapper.py`) with path to the socket and the handler to run. The Python
+wrapper connects to this socket upon startup.
 
-# Wire Protocol
 The wite protocol is line oriented where every line is a JSON object.
 - Go sends events (encoded using `EventJSONEncoder`)
-- RPC cleint sends lines starting with letter specifying type and then JSON object
+- Python sends lines starting with letter specifying type and then JSON object
     - 'r' Handler reply
     - 'l' Log messages
 	- 'm' Metric messages
-
-# Event Encoding
-- Body is encoded in base64 (to allow binary data)
-- Timestamp is seconds since epoch
 */
 
-// Package rpc implmenets Python runtime
-package rpc
+// Package python implmenets Python runtime
+package python
