+++ pkg/processor/runtime/python/runtime.go
@@ -17,39 +17,145 @@ limitations under the License.
 package python
 
 import (
+	"bufio"
+	"encoding/base64"
+	"encoding/json"
 	"fmt"
+	"net"
 	"os"
 	"os/exec"
 	"strings"
+	"time"
 
 	"github.com/nuclio/nuclio/pkg/common"
 	"github.com/nuclio/nuclio/pkg/errors"
 	"github.com/nuclio/nuclio/pkg/processor/runtime"
-	"github.com/nuclio/nuclio/pkg/processor/runtime/rpc"
 
 	"github.com/nuclio/nuclio-sdk"
+	"github.com/rs/xid"
 )
 
+// TODO: Find a better place (both on file system and configuration)
+const (
+	socketPathTemplate = "/tmp/nuclio-py-%s.sock"
+	connectionTimeout  = 10 * time.Second
+	eventTimeout       = 5 * time.Minute
+)
+
+type result struct {
+	StatusCode   int                    `json:"status_code"`
+	ContentType  string                 `json:"content_type"`
+	Body         string                 `json:"body"`
+	BodyEncoding string                 `json:"body_encoding"`
+	Headers      map[string]interface{} `json:"headers"`
+
+	DecodedBody []byte
+	err         error
+}
+
 type python struct {
-	*rpc.Runtime
+	runtime.AbstractRuntime
 	configuration *runtime.Configuration
+	eventEncoder  *EventJSONEncoder
+	outReader     *bufio.Reader
+	socketPath    string
+}
+
+type pythonLogRecord struct {
+	DateTime string                 `json:"datetime"`
+	Level    string                 `json:"level"`
+	Message  string                 `json:"message"`
+	With     map[string]interface{} `json:"with"`
 }
 
 // NewRuntime returns a new Python runtime
 func NewRuntime(parentLogger nuclio.Logger, configuration *runtime.Configuration) (runtime.Runtime, error) {
 	logger := parentLogger.GetChild("python")
-	newPythonRuntime := &python{
-		configuration: configuration,
-	}
-	newPythonRuntime.Logger = logger // We *must* initialize logger here
 
 	var err error
-	newPythonRuntime.Runtime, err = rpc.NewRPCRuntime(logger, configuration, newPythonRuntime.runWrapper)
 
-	return newPythonRuntime, err
+	abstractRuntime, err := runtime.NewAbstractRuntime(logger, configuration)
+	if err != nil {
+		return nil, errors.Wrap(err, "Can't create AbstractRuntime")
+	}
+
+	newPythonRuntime := &python{
+		AbstractRuntime: *abstractRuntime,
+		configuration:   configuration,
+	}
+
+	// create socket path
+	newPythonRuntime.socketPath = newPythonRuntime.createSocketPath()
+
+	listener, err := newPythonRuntime.createListener()
+	if err != nil {
+		return nil, errors.Wrapf(err, "Can't listen on %q", newPythonRuntime.socketPath)
+	}
+
+	if err = newPythonRuntime.runWrapper(); err != nil {
+		return nil, errors.Wrap(err, "Can't run wrapper")
+	}
+
+	unixListener, ok := listener.(*net.UnixListener)
+	if !ok {
+		return nil, errors.Wrap(err, "Can't get underlying Unix listener")
+	}
+	if err = unixListener.SetDeadline(time.Now().Add(connectionTimeout)); err != nil {
+		return nil, errors.Wrap(err, "Can't set deadline")
+	}
+	conn, err := listener.Accept()
+	if err != nil {
+		return nil, errors.Wrap(err, "Can't get connection from Python wrapper")
+	}
+
+	newPythonRuntime.eventEncoder = NewEventJSONEncoder(newPythonRuntime.Logger, conn)
+	newPythonRuntime.outReader = bufio.NewReader(conn)
+
+	return newPythonRuntime, nil
 }
 
-func (py *python) runWrapper(socketPath string) error {
+func (py *python) ProcessEvent(event nuclio.Event, functionLogger nuclio.Logger) (interface{}, error) {
+	py.Logger.DebugWith("Processing event",
+		"name", py.configuration.Meta.Name,
+		"version", py.configuration.Spec.Version,
+		"eventID", event.GetID())
+
+	resultChan := make(chan *result)
+	go py.handleEvent(functionLogger, event, resultChan)
+
+	select {
+	case result := <-resultChan:
+		py.Logger.DebugWith("Python executed",
+			"status", result.StatusCode,
+			"eventID", event.GetID())
+		return nuclio.Response{
+			Body:        result.DecodedBody,
+			ContentType: result.ContentType,
+			Headers:     result.Headers,
+			StatusCode:  result.StatusCode,
+		}, nil
+	case <-time.After(eventTimeout):
+		return nil, fmt.Errorf("handler timeout after %s", eventTimeout)
+	}
+}
+
+func (py *python) createListener() (net.Listener, error) {
+	if common.FileExists(py.socketPath) {
+		if err := os.Remove(py.socketPath); err != nil {
+			return nil, errors.Wrapf(err, "Can't remove socket at %q", py.socketPath)
+		}
+	}
+
+	py.Logger.DebugWith("Creating listener socket", "path", py.socketPath)
+
+	return net.Listen("unix", py.socketPath)
+}
+
+func (py *python) createSocketPath() string {
+	return fmt.Sprintf(socketPathTemplate, xid.New().String())
+}
+
+func (py *python) runWrapper() error {
 	wrapperScriptPath := py.getWrapperScriptPath()
 	py.Logger.DebugWith("Using Python wrapper script path", "path", wrapperScriptPath)
 	if !common.IsFile(wrapperScriptPath) {
@@ -76,7 +182,7 @@ func (py *python) runWrapper(socketPath string) error {
 	args := []string{
 		pythonExePath, wrapperScriptPath,
 		"--handler", handler,
-		"--socket-path", socketPath,
+		"--socket-path", py.socketPath,
 	}
 
 	py.Logger.DebugWith("Running wrapper", "command", strings.Join(args, " "))
@@ -89,11 +195,117 @@ func (py *python) runWrapper(socketPath string) error {
 	return cmd.Start()
 }
 
+func (py *python) handleEvent(functionLogger nuclio.Logger, event nuclio.Event, resultChan chan *result) {
+	unmarshalledResult := &result{}
+
+	// Send event
+	if unmarshalledResult.err = py.eventEncoder.Encode(event); unmarshalledResult.err != nil {
+		resultChan <- unmarshalledResult
+		return
+	}
+
+	var data []byte
+
+	// Read logs & output
+	for {
+		data, unmarshalledResult.err = py.outReader.ReadBytes('\n')
+
+		if unmarshalledResult.err != nil {
+			py.Logger.WarnWith("Failed to read from connection", "err", unmarshalledResult.err)
+
+			resultChan <- unmarshalledResult
+			return
+		}
+
+		switch data[0] {
+		case 'r':
+
+			// try to unmarshall the result
+			if unmarshalledResult.err = json.Unmarshal(data[1:], unmarshalledResult); unmarshalledResult.err != nil {
+				resultChan <- unmarshalledResult
+				return
+			}
+
+			switch unmarshalledResult.BodyEncoding {
+			case "text":
+				unmarshalledResult.DecodedBody = []byte(unmarshalledResult.Body)
+			case "base64":
+				unmarshalledResult.DecodedBody, unmarshalledResult.err = base64.StdEncoding.DecodeString(unmarshalledResult.Body)
+			default:
+				unmarshalledResult.err = fmt.Errorf("Unknown body encoding - %q", unmarshalledResult.BodyEncoding)
+			}
+
+			// write back to result channel
+			resultChan <- unmarshalledResult
+
+			return // reply is the last message the python wrapper sends
+
+		case 'm':
+			py.handleReponseMetric(functionLogger, data[1:])
+
+		case 'l':
+			py.handleResponseLog(functionLogger, data[1:])
+		}
+	}
+}
+
+func (py *python) handleResponseLog(functionLogger nuclio.Logger, response []byte) {
+	var logRecord pythonLogRecord
+
+	if err := json.Unmarshal(response, &logRecord); err != nil {
+		py.Logger.ErrorWith("Can't decode log", "error", err)
+		return
+	}
+
+	logger := py.resolveFunctionLogger(functionLogger)
+	logFunc := logger.DebugWith
+
+	switch logRecord.Level {
+	case "error", "critical", "fatal":
+		logFunc = logger.ErrorWith
+	case "warning":
+		logFunc = logger.WarnWith
+	case "info":
+		logFunc = logger.InfoWith
+	}
+
+	vars := common.MapToSlice(logRecord.With)
+	logFunc(logRecord.Message, vars...)
+}
+
+func (py *python) handleReponseMetric(functionLogger nuclio.Logger, response []byte) {
+	var metrics struct {
+		DurationSec float64 `json:"duration"`
+	}
+
+	logger := py.resolveFunctionLogger(functionLogger)
+	if err := json.Unmarshal(response, &metrics); err != nil {
+		logger.ErrorWith("Can't decode metric", "error", err)
+		return
+	}
+
+	if metrics.DurationSec == 0 {
+		logger.ErrorWith("No duration in metrics", "metrics", metrics)
+		return
+	}
+
+	py.Statistics.DurationMilliSecondsCount++
+	py.Statistics.DurationMilliSecondsSum += uint64(metrics.DurationSec * 1000)
+}
+
 func (py *python) getEnvFromConfiguration() []string {
 	return []string{
 		fmt.Sprintf("NUCLIO_FUNCTION_NAME=%s", py.configuration.Meta.Name),
 		fmt.Sprintf("NUCLIO_FUNCTION_DESCRIPTION=%s", py.configuration.Spec.Description),
 		fmt.Sprintf("NUCLIO_FUNCTION_VERSION=%d", py.configuration.Spec.Version),
+	}
+}
+
+func (py *python) getEnvFromEvent(event nuclio.Event) []string {
+	return []string{
+		fmt.Sprintf("NUCLIO_EVENT_ID=%s", event.GetID()),
+		fmt.Sprintf("NUCLIO_EVENT_SOURCE_CLASS=%s", event.GetSource().GetClass()),
+		fmt.Sprintf("NUCLIO_EVENT_SOURCE_KIND=%s", event.GetSource().GetKind()),
 	}
 }
 
@@ -143,4 +355,12 @@ func (py *python) getPythonExePath() (string, error) {
 	}
 
 	return "", errors.Wrap(err, "Can't find python executable")
+}
+
+// resolveFunctionLogger return either functionLogger if provided or root logger if not
+func (py *python) resolveFunctionLogger(functionLogger nuclio.Logger) nuclio.Logger {
+	if functionLogger == nil {
+		return py.Logger
+	}
+	return functionLogger
 }
