+++ pkg/processor/trigger/http/test/suite/suite.go
@@ -89,7 +89,6 @@ func (suite *TestSuite) DeployFunctionAndRequest(deployOptions *platform.DeployO
 	}
 
 	return suite.DeployFunction(deployOptions, func(deployResult *platform.DeployResult) bool {
-		suite.WaitForContainer(deployResult.Port)
 
 		// modify request port to that of the deployed
 		request.RequestPort = deployResult.Port
