--- test/_functions/common/json-parser-with-inline-function-config/nodejs/handler.js
@@ -1,35 +1,0 @@
-/*
-Copyright 2017 The Nuclio Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-// @nuclio.configure
-//
-// function.yaml:
-//   spec:
-//     runtime: nodejs
-//     handler: handler
-//     build:
-//       commands:
-//       - npm install -g moment
-//
-
-// Check that npm install worked
-var moment = require('moment');
-
-exports.handler = function(context, event) {
-    context.log_info('Hello from NodeJS');
-    var body = JSON.parse(event.body.toString());
-    context.callback(body.return_this);
-}
